requirements.txt
streamlit
tensorflow==2.20.0
pillow
h5py
numpy
pandas
plotly